+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- DEP: drop support for Python 3.9, require 3.10 or newer
+
 ## 0.4.0 - 2025-07-10
 
 - DOC: consistently indicate argument types (positional-only and keyword-only)